requirements.txt
# Data and ML tools
numpy<2
pandas==2.2.3
scikit-learn==1.4.0
tqdm==4.66.1
ipykernel==6.29.4


# Image processing
opencv-python==4.9.0.80
Pillow==10.3.0
rembg==2.0.50

# PyTorch (CUDA 11.8)
torch==2.2.1+cu118
torchvision==0.17.1+cu118

# ONNX for production (GPU support)
onnx==1.14.1
onnxruntime-gpu==1.17.0

# Include PyTorch wheel source for CUDA versions
--find-links https://download.pytorch.org/whl/torch_stable.html